+++ CHANGELOG.md
@@ -2,8 +2,11 @@ hook_update_deploy_tools 7.x-1.x **-**-****
 -----------------------------------------------
 * [New items go at the top of this list.]
 
-hook_update_deploy_tools 7.x-1.33 Nov 09, 2017
+
+hook_update_deploy_tools 7.x-1.34 Nov 09, 2017
 -----------------------------------------------
+* Issue #2922294: Modules::uninstall improve handling dependencies
+  https://www.drupal.org/node/2922294
 * bugfix Modules::enable Incorrect passing of module name.
 
 hook_update_deploy_tools 7.x-1.32 Nov 08, 2017